ixp4xx_eth: move mdio bus code into a separate file
This should really be a separate platform_driver. As a first
step, move the contents into a new file, but don't change the
way the driver works just yet.

The only functional change is to check for the phy first
so we can defer probing to get the two pieces in the right
order.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>