watchdog: ixp4xx: move mmio access to timer driver
The registers for watchdog and timer overlap, and the timer
driver already has a way to get hold of them, which the watchdog
does not.

Move the low-level access for the timeout and the boot status
into the clocksource driver to make the watchdog driver more
portable and enable compile testing it.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>