ARM: samsung: make pm-debug platform independent
The pm-debug code is one of the few things shared between s3c24xx/s3c64xx
and the newer s5pv210. In order to make s5pv210 independent of plat-samsung,
change the common bits of this code to no longer reference the s3c specific
bits.

In particular, all the CPU checks need to be moved out of the
common code into platform specific files.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>