[SUBMITTED 20190906] lz4: make LZ4HC_setExternalDict as non-static
kbuild warns for exported static symbols. This one seems to
be meant as an external API but does not have any in-kernel
users:

WARNING: "LZ4HC_setExternalDict" [vmlinux] is a static EXPORT_SYMBOL

I suppose the function should not just get removed since it would
be nice to stay close to the upstream version of lz4hc, so just
make it global.

The modpost commit is correct, but is listed here as the one
that introduced the warning.

Fixes: 15bfc2348d54 ("modpost: check for static EXPORT_SYMBOL* functions")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>