video: backlight: tosa: use gpio lookup table
The driver should not require a machine specific header. Change
it to pass the gpio line through a lookup table, and move the
timing generator definitions into the drivers itself.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I'm not overly confident that I got the correct device names
for the lookup table, it would be good if someone could
double-check.