ARM: pxa: disable mtd-xip support
Supporting this with a multiplatform configuration is a bit tricky,
but could be done if necessary. If nobody is actually using it though,
it can simply be removed.

None of the defconfigs use it today, and compilation was broken between
2011 and 2017 before I fixed it as part of randconfig testing, so I think
nobody will miss it. This leaves sa1100 as the only platform with mtd-xip
support: omap1 xip support was broken immediately after it got merged,
and on i.MX it got removed after it turned out to be broken as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>