ARM: omap: move dma_global_context code to mach-omap2
omap34xx calls omap_dma_global_context_save/restore in the suspend/resume
code, which is always built-in, but there are no other callers.

As preparation for moving the rest of the omap-dma code into the driver,
move the global context handling into omap2 itself, which simplifies
the code a bit.

Unfortunately, that code is unaware of which channels are actually
in use, so instead of only clearing the active channels, it now has
to clear all of them.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>