crypto: chacha - move existing library code into lib/crypto
Move the existing shared ChaCha code into lib/crypto, and at the
same time, split the support header into a public version, and an
internal version that is only intended for consumption by crypto
implementations.

At the same time, refactor the generic implementation so it only gets
exposed as the chacha_crypt() library function if the architecture does
not override it with its own implementation, potentially falling back
to the generic routine if needed.

And while at it, tidy up lib/crypto/Makefile a bit so we are ready for
some new arrivals.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>