amdgpu: dc: wrap float types
The use of hardfloat types makes the dc driver nonportable. As a first
step to avoiding this, change all uses of 'float' and 'double' to
new 'amdgpu_dc_float' and 'amdgpu_dc_double' typedefs.

The names are meant to be long enough to make it undesirable to use
them more than absolutely necessary, especially spreading them
outside of the dc driver.

As there are only typedefs to the original types, this patch should
not change the generated code.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>