af_packet: stop using getnstimeofday
For some reason, this file has escaped all previous cleanup that removed
the use of 'struct timespec' and associated interfaces, and is now the
last user of getnstimeofday(), and it is one that would suffer from the
y2038 overflow in theory.

Change it to use the modern ktime interfaces so we can remove
getnstimeofday, and add a comment about the fact that this will overflow
eventually.

As the timestamp fields are already unsigned 32-bit numbers, this can
work beyond y2038 as long as user space interprets it as an unsigned
number consistently, or casts it to a 64-bit time_t.

If we ever get a v4 interface, this should use a format that avoids the
overflow, such as a 64-bit nanosecond counter.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>