nds32: open-code timespec_add_ns()
The nds32 vdso is now the last user of the deprecated timespec_add_ns().

Change it to an open-coded version like the one it already uses in
do_realtime(). What we should really do though is to use the
generic vdso implementation that is now used in x86. arm and mips.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>