time_types: add __kernel_old_timespec
The 'struct timespec' definition can no longer be part of the
uapi headers because it conflicts with a a now incompatible libc
definition. Also, we really want to remove it in order to prevent
new uses from creeping in.

Add a replacement __kernel_old_timespec structure along the lines
of __kernel_old_timeval.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>