[SUBMITTED 20191209] [net-next] wan: remove old frame relay driver
The DLCI framework and SDLA driver are as obsolete as it gets,
this is an ISA add-on card from the mid-1990s that has only
seen sporadic cleanups.

The website listed for downloading the user space tools has dropped
off the internet at some point around 2010, but Debian already
dropped the packages ten years earlier when they no longer built
correctly.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>