MAINTAINERS: add new X.25 maintainer
Martin Schiller is using the Linux X.25 stack on top of HDLC and
X.21 networks. He agreed to be listed as a maintainer to take
care of odd fixes.

Add him as the primary contact for net/x25 and net/lapb, as well
as a reviewer for drivers/net/wan, which contains the HDLC code.

Cc: Martin Schiller <ms@dev.tdt.de>
Cc: Andrew Hendry <andrew.hendry@gmail.com>
Cc: Krzysztof Halasa <khc@pm.waw.pl>
Link: https://lore.kernel.org/netdev/407acd92c92c3ba04578da89b1a0f191@dev.tdt.de/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>