arch: add global CONFIG_PAGE_SIZE_{4,8,16,32,64,256}K symbols
Configuration of page sizes is a mess, and any code that only works
on certain sizes has to know about all the architectures and their
own symbols.

Add a global selection of the page size to consolidate this, leaving the
architecture specific options around for the moment, those can be replaced
with ARCH_HAVE_PAGE_SIZE_4KB etc. symbols later on, one architecture at
a time.

On Hexagon, MIPS, and SuperH, we already use the common symbols, so
for these the Kconfig file has to be changed already to avoid symbol
clashes. The architectures that do not support 4KB pages (alpha, sparc64,
openrisc and some m68k) select ARCH_MIN_PAGE_SIZE_8K for simplification.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>