drm: decouple from CONFIG_FB
CONFIG_DRM_KMS_FB_HELPER selects CONFIG_FB, which is something it
really should not, to avoid circular dependencies and accidentally
including potentially dangerous user interfaces in the kernel,
so change this into a 'depends on' check.

Two device drivers currently select CONFIG_DRM_KMS_FB_HELPER, but
as far as I can tell, they do not really need to any more, so those
selects can be removed.

This leaves DRM_FBDEV_EMULATION as the only thing that rightfully
selects CONFIG_DRM_KMS_FB_HELPER, and this now has to depend on
CONFIG_FB and its dependencies.
When CONFIG_FB is a loadable module, DRM_KMS_FB_HELPER cannot be
part of the built-in subsystem, so add dependency to ensure this
can only be enabled if the DRM module can successfully be linked.

Reviewed-by: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>