iio: adc: ad7476: fix clang -Wpointer-bool-conversion warning
Checking the pointer value of st->chip_info->convst_channel is pointless
since this this an array inside of a struct: even if st->chip_info is NULL,
the pointer is non-zero. Clang warns about this:

drivers/iio/adc/ad7476.c:312:40: warning: address of array 'st->chip_info->convst_channel' will always evaluate to 'true' [-Wpointer-bool-conversion]
        if (st->convst_gpio && st->chip_info->convst_channel)
                            ~~ ~~~~~~~~~~~~~~~^~~~~~~~~~~~~~

I could not come up with a sane way to check whether the entry
is valid, so just remove the check and keep the behavior as it
is today but without the warning.

Fixes: 3a6af93dd66e ("iio: adc: ad7476: Add IIO_CHAN_INFO_RAW for AD7091R")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>